+++ lambdas/manifest_indexer/src/t4_lambda_manifest_indexer/__init__.py
@@ -75,7 +75,7 @@ def _get_metadata_fields(path: tuple, d: dict):
                     type_ = "keyword" if len(v) <= MAX_KEYWORD_LEN else "text"
             elif isinstance(v, bool):
                 type_ = "boolean"
-            elif isinstance(v, (int, float)):
+            elif isinstance(v, int | float):
                 # XXX: do something on ints that can't be converted to float without loss?
                 type_ = "double"
             elif isinstance(v, list):